refactor: pool component names
